updated DI modules

+++ backend/src/api/data-access-di/blocks/unsubscribe.js
@@ -5,6 +5,11 @@ const unsubscribe = async data => {
 
   data = JSON.parse(data);
 
+  if (data.follower.length == 0 || data.followed.length == 0) {
+    log('Invalid input data', 'err:unsubscribe');
+    return feedback;
+  }
+
   await relSubscriptions.deleteOne( data )
   .then(() => {
     log('%s unsubscribed from %s', 'unsubscribe', data.follower, data.followed);
